update frameworks, fix some resolves, start consumer
