+++ src/Application.CommandHandlers/Application.CommandHandlers.csproj
@@ -8,9 +8,9 @@
 
   <ItemGroup>
     <PackageReference Include="PetProjects.Framework.Cqrs" Version="1.0.17337.3" />
-    <PackageReference Include="PetProjects.Framework.Kafka" Version="0.1.17344.3-beta" />
+    <PackageReference Include="PetProjects.Framework.Kafka" Version="0.1.17344.2" />
     <PackageReference Include="PetProjects.Framework.Logging.Contracts" Version="0.9.17343.4" />
-    <PackageReference Include="PetProjects.MicroTransactions.Events" Version="0.1.17344.1" />
+    <PackageReference Include="PetProjects.MicroTransactions.Events" Version="0.1.17344.2" />
   </ItemGroup>
 
   <ItemGroup>
