+++ src/Application.Consumers/Application.Consumers.csproj
@@ -8,8 +8,8 @@
 
   <ItemGroup>
     <PackageReference Include="PetProjects.Framework.Cqrs" Version="1.0.17337.3" />
-    <PackageReference Include="PetProjects.Framework.Kafka" Version="0.1.17344.3-beta" />
-    <PackageReference Include="PetProjects.MicroTransactions.Commands" Version="0.1.17344.1" />
+    <PackageReference Include="PetProjects.Framework.Kafka" Version="0.1.17344.2" />
+    <PackageReference Include="PetProjects.MicroTransactions.Commands" Version="0.1.17344.2" />
   </ItemGroup>
 
   <ItemGroup>
