+++ src/Infrastructure.Configurations/Infrastructure.Configurations.csproj
@@ -8,15 +8,15 @@
 
   <ItemGroup>
     <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="2.0.0" />
+    <PackageReference Include="Microsoft.Extensions.Logging.Console" Version="2.0.0" />
     <PackageReference Include="Newtonsoft.Json" Version="10.0.3" />
     <PackageReference Include="PetProjects.Framework.Consul" Version="0.9.17330.3" />
     <PackageReference Include="PetProjects.Framework.Cqrs" Version="1.0.17337.3" />
     <PackageReference Include="PetProjects.Framework.Cqrs.Extensions.AspNetCore" Version="1.0.17337.3" />
-    <PackageReference Include="PetProjects.Framework.Kafka" Version="0.1.17344.3-beta" />
     <PackageReference Include="PetProjects.Framework.Logging.Contracts" Version="0.9.17343.4" />
     <PackageReference Include="PetProjects.Framework.Logging.Producer" Version="0.9.17343.4" />
-    <PackageReference Include="PetProjects.MicroTransactions.Commands" Version="0.1.17344.1" />
-    <PackageReference Include="PetProjects.MicroTransactions.Events" Version="0.1.17344.1" />
+    <PackageReference Include="PetProjects.MicroTransactions.Commands" Version="0.1.17344.2" />
+    <PackageReference Include="PetProjects.MicroTransactions.Events" Version="0.1.17344.2" />
   </ItemGroup>
 
   <ItemGroup>
