+++ test/integration/Integration.Tests/Integration.Tests/Integration.Tests.csproj
@@ -12,4 +12,9 @@
     <PackageReference Include="xunit.runner.visualstudio" Version="2.2.0" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="CassandraDb\" />
+    <Folder Include="Producer\" />
+  </ItemGroup>
+
 </Project>
